server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pyyaml>=6.0
mysql-connector-python>=8.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
fastapi-mail>=1.4.0
pydantic-settings>=2.2.0

# Development tools
ruff>=0.3.0
black>=24.2.0
pytest>=8.0.0
iterative-stratification>=0.1.7
